Change variable types to use their proper RECT type

+++ include/fft/event_card.h
@@ -38,7 +38,7 @@ void card_io_consume_hardware_events(void);
 s32 card_io_create_new_with_retries(s32 port, s32 retry_limit);
 s32 card_io_format_selected_slot(void);
 s32 card_io_get_info_with_retries(s32 port, s32 retry_limit);
-u8 card_io_get_selected_slot(void);
+s32 card_io_get_selected_slot(void);
 s32 card_io_load_with_retries(s32 port, s32 retry_limit);
 s32 card_io_open_or_create_selected_card(void);
 s32 card_io_poll_bios_event_result(void);
